+++ packages/core/CHANGELOG.md
@@ -1,2 +1,5 @@
+## 1.0.1-1.2.2
+- Bugfixes
+- Remove `serviceworker` option
 ## 1.0.0
 - It exists